// apps/scm/src/modules-features/ModuleThanhLapHoiDongTuVanTuyenChon/interfaces/IThanhLapHoiDongTuVanTuyenChonViewModal.ts
export interface IThanhLapHoiDongTuVanTuyenChonViewModal {
    code?: string;
    tenHoiDong?: string;
    ngayHop?: Date;
    thoiGianHop?: string;
    diaDiemHop?: string;
    chuTichHoiDong?: string;
    thuKyHoiDong?: string;
    danhSachThanhVien?: string;
    dangKyXetDuyet?: string;
    trangThaiHoiDong?: string;
    file?: File;
}